Configure Azure SDK clients with RateLimitCircuitBreaker
