+++ client/senddecorator/ratelimit_circuitbreaker.go
@@ -17,6 +17,11 @@ const (
 	defaultWaitAfterTooManyRequests = 6 * time.Minute
 )
 
+func init() {
+	// ONE DOES NOT SIMPLY RETRY ON HTTP 429.
+	autorest.StatusCodesForRetry = removeElementFromSlice(autorest.StatusCodesForRetry, http.StatusTooManyRequests)
+}
+
 // RateLimitCircuitBreaker utilizes simple gatekeeper implementation to hold
 // off from making any additional requests when server responds HTTP 429 Too
 // Many Requests.
@@ -47,4 +52,19 @@ func RateLimitCircuitBreaker(g *gatekeeper.Gatekeeper) autorest.SendDecorator {
 			return resp, err
 		})
 	}
+}
+
+func removeElementFromSlice(xs []int, x int) []int {
+	for i, v := range xs {
+		if v == x {
+			// Shift end of slice to the left by one.
+			copy(xs[i:], xs[i+1:])
+			// Truncate the last element.
+			xs = xs[:len(xs)-1]
+			// Call it a day.
+			break
+		}
+	}
+
+	return xs
 }
